little integration test refactor
