+++ test/integration/mock/assessment.ts
@@ -1,4 +1,4 @@
-import { AssessmentName } from '@meta/assessment'
+import { AssessmentName, OriginalDataPoint } from '@meta/assessment'
 
 export const assessmentParams = {
   props: {
@@ -11,4 +11,4 @@ export const assessmentCycleName = '2020'
 export const originalDataPoint = {
   countryIso: 'AFG',
   year: '2019',
-}
+} as OriginalDataPoint
